new dbmanager wip
